requirements.txt
flask
flask_restful
flask_cors

datamodel-code-generator
pydantic

python-dotenv
openai
langchain
tiktoken